+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - general cleanups and linting fixes - that includes several renames of packagenames and types.
 - cart module:
     - Has a new secondary port: PlaceOrderService
+    - The meaning of DeliveryInfo.Method has changed! The former meaning is now represented in the property DeliveryInfo.Workflow. See Readme of cart ackage for details